refactor scatter3d.go
